restructure(redux): Move thread actions into thread folder
