+++ src/redux/actions/thread/destroyThread.js
@@ -1,0 +1,9 @@
+import * as types from 'redux/types';
+
+export default function destroyThread () {
+    return dispatch => {
+        dispatch({
+            type: types.THREAD_DESTROYED
+        });
+    }
+}
