+++ src/redux/actions/thread/index.js
@@ -1,0 +1,4 @@
+export {default as destroyThread} from './destroyThread';
+export {default as monitorThread} from './monitorThread';
+export {default as unmonitorThread} from './unmonitorThread';
+export {default as updateMonitoredThread} from './updateMonitoredThread';
